Finally i got uppercase Hex digits!
Turns out if you dont set the parserOptions eslint seems to fail
silently in the background compared to when run in the terminal.
I think it doesn't recognize the 'import' statement. Or at least
that's what the failure in the cli was.
Anyway as soon as i added the parserOptions it worked.
Fingers crossed it stays that way.
Just fyi remember not to add:
```js
    "ecmaFeatures": {
      "jsx": true
    }
```
That actually stopped it running on the Emulator.ts file.
Not sure why i'll figure it out. Just tired right now.